requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
youtube_dl==2021.12.17
ffmpeg-python==0.2.0
PyNaCl==1.5.0
yt-dlp==2024.08.06